managed to carry player's state from one leveel to the next

+++ source/core/src/main/com/deco2800/game/entities/configs/PlayerConfig.java
@@ -15,7 +15,7 @@ public class PlayerConfig extends BaseEntityConfig  {
   public int health;
   public int ammo;
   public int bandages;
-  public int currentGameLevel;
+  public double currentGameLevel;
   public String ability;
   public String meleeFilePath;
   public String meleeWeaponType;
